{"schema":"decodex/commit/1","summary":"Split Radar CLI and queue modules","authority":"manual"}
